# backend/src/app/generated/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@example.com
    name: API Support
  description: "A simple note-taking application API that allows anonymous browsing\
    \ of latest public notes \nand authenticated access to personal plain text notebook\
    \ features.\n\n## Authentication\n- Public endpoints require no authentication\n\
    - Private endpoints use Firebase Auth with session cookies\n- Anonymous users\
    \ are automatically created for notebook access\n- Regular users can access basic\
    \ profile management features\n"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Simple Note Application API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.noteapp.com
- description: Staging server
  url: https://staging-api.noteapp.com
- description: Local development server
  url: http://dev-api.noteapp.com
security:
- SessionAuth: []
- BearerAuth: []
- {}
...
paths:
...
  /me:
    get:
      description: |
        Get current user's profile information (regular users only).
        Anonymous users will receive a 403 error.
      operationId: getUserProfile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
          description: User profile information
        "403":
          content:
            application/json:
              example:
                status: error
                error:
                  code: ANONYMOUS_PROFILE_FORBIDDEN
                  message: Anonymous users do not have profiles. Please create an
                    account to access profile features.
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Anonymous users cannot access profile
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
      security:
      - SessionAuth: []
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Profile
    patch:
      description: |
        Update current user's profile information (regular users only).
        Anonymous users will receive a 403 error.
      operationId: updateUserProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
          description: Profile updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request data
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Anonymous users cannot update profile
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authentication required
      security:
      - SessionAuth: []
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User Profile
...
components:
  schemas:
...
    UserProfile:
      description: User profile information (available for regular users only)
      example:
        settings:
          editor: split
          theme: dark
          locale: en-US
        createdAt: 2023-11-15T09:00:00Z
        isAnonymous: false
        avatarUrl: https://example.com/avatars/jane.jpg
        displayName: Jane Developer
        userId: user_789ghi
        updatedAt: 2023-12-01T12:00:00Z
      properties:
        userId:
          description: Firebase UID of the user
          example: user_789ghi
          title: userId
          type: string
        displayName:
          description: User's display name
          example: Jane Developer
          nullable: true
          title: displayName
          type: string
        avatarUrl:
          description: URL to user's avatar image
          example: https://example.com/avatars/jane.jpg
          nullable: true
          title: avatarUrl
          type: string
        settings:
          $ref: "#/components/schemas/UserProfile_settings"
        isAnonymous:
          description: Whether this is an anonymous user (always false for profiles)
          example: false
          title: isAnonymous
          type: boolean
        createdAt:
          description: When the user account was created
          example: 2023-11-15T09:00:00Z
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          description: When the profile was last updated
          example: 2023-12-01T12:00:00Z
          format: date-time
          title: updatedAt
          type: string
      required:
      - createdAt
      - isAnonymous
      - settings
      - updatedAt
      - userId
      title: UserProfile
      type: object
...
    UserProfileResponse:
      $ref: "#/components/schemas/user-profile-response"
...
    ErrorResponse:
      $ref: "#/components/schemas/error-response"
...
    UpdateUserProfileRequest:
      $ref: "#/components/schemas/update-user-profile-request"
...
    ApiResponseStatus:
      description: Status of the API response
      enum:
      - success
      - error
      example: success
      title: ApiResponseStatus
      type: string
    ErrorDetail:
      description: Detailed error information
      example:
        code: INVALID_REQUEST
        field: title
        message: The provided data is invalid
      properties:
        code:
          description: Error code for programmatic handling
          example: INVALID_REQUEST
          title: code
          type: string
        message:
          description: Human-readable error message
          example: The provided data is invalid
          title: message
          type: string
        field:
          description: Field name that caused the error (if applicable)
          example: title
          nullable: true
          title: field
          type: string
      required:
      - code
      - message
      title: ErrorDetail
      type: object
...
    error-response:
      description: Error response wrapper
      example:
        error:
          code: INVALID_REQUEST
          field: title
          message: The provided data is invalid
        status: success
      properties:
        status:
          $ref: "#/components/schemas/ApiResponseStatus"
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
      - error
      - status
      title: error-response
      type: object
...
    user-profile-response:
      description: Response wrapper for user profile
      example:
        data:
          settings:
            editor: split
            theme: dark
            locale: en-US
          createdAt: 2023-11-15T09:00:00Z
          isAnonymous: false
          avatarUrl: https://example.com/avatars/jane.jpg
          displayName: Jane Developer
          userId: user_789ghi
          updatedAt: 2023-12-01T12:00:00Z
        status: success
      properties:
        status:
          $ref: "#/components/schemas/ApiResponseStatus"
        data:
          $ref: "#/components/schemas/UserProfile"
      required:
      - data
      - status
      title: user-profile-response
      type: object
    update-user-profile-request:
      description: Request payload for updating user profile
      properties:
        displayName:
          description: Updated display name
          example: Jane Smith
          maxLength: 100
          minLength: 1
          nullable: true
          title: displayName
          type: string
        avatarUrl:
          description: Updated avatar image URL
          example: https://example.com/avatars/jane-new.jpg
          format: uri
          nullable: true
          title: avatarUrl
          type: string
        settings:
          $ref: "#/components/schemas/update_user_profile_request_settings"
      title: update-user-profile-request
      type: object
...
    UserProfile_settings:
      description: User preferences and settings
      example:
        editor: split
        theme: dark
        locale: en-US
      properties:
        locale:
          description: User's preferred locale
          example: en-US
          title: locale
          type: string
        theme:
          description: User's preferred theme
          enum:
          - light
          - dark
          - system
          example: dark
          title: theme
          type: string
        editor:
          description: User's preferred editor mode
          enum:
          - rich
          - markdown
          - split
          example: split
          title: editor
          type: string
      required:
      - editor
      - locale
      - theme
      title: UserProfile_settings
      type: object
...
    update_user_profile_request_settings:
      description: Updated user settings
      properties:
        locale:
          description: User's preferred locale
          example: en-US
          title: locale
          type: string
        theme:
          description: User's preferred theme
          enum:
          - light
          - dark
          - system
          example: light
          title: theme
          type: string
        editor:
          description: User's preferred editor mode
          enum:
          - rich
          - markdown
          - split
          example: rich
          title: editor
          type: string
      title: update_user_profile_request_settings
      type: object
...
  securitySchemes:
    SessionAuth:
      description: Firebase session cookie
      in: cookie
      name: __session
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      description: Firebase ID token
      scheme: bearer
      type: http