# backend/src/app/generated/src/generated_fastapi_server/models/api_response_status.py
# coding: utf-8

"""
    Simple Note Application API

    A simple note-taking application API that allows anonymous browsing of latest public notes  and authenticated access to personal plain text notebook features.  ## Authentication - Public endpoints require no authentication - Private endpoints use Firebase Auth with session cookies - Anonymous users are automatically created for notebook access - Regular users can access basic profile management features 

    The version of the OpenAPI document: 1.0.0
    Contact: support@example.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class ApiResponseStatus(str, Enum):
    """
    Status of the API response
    """

    """
    allowed enum values
    """
    SUCCESS = 'success'
    ERROR = 'error'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ApiResponseStatus from a JSON string"""
        return cls(json.loads(json_str))