# backend/src/app/generated/src/generated_fastapi_server/models/error_detail.py
# coding: utf-8

"""
    Simple Note Application API

    A simple note-taking application API that allows anonymous browsing of latest public notes  and authenticated access to personal plain text notebook features.  ## Authentication - Public endpoints require no authentication - Private endpoints use Firebase Auth with session cookies - Anonymous users are automatically created for notebook access - Regular users can access basic profile management features 

    The version of the OpenAPI document: 1.0.0
    Contact: support@example.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class ErrorDetail(BaseModel):
    """
    Detailed error information
    """ # noqa: E501
    code: StrictStr = Field(description="Error code for programmatic handling")
    message: StrictStr = Field(description="Human-readable error message")
    var_field: Optional[StrictStr] = Field(default=None, description="Field name that caused the error (if applicable)", alias="field")
    __properties: ClassVar[List[str]] = ["code", "message", "field"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of ErrorDetail from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # set to None if var_field (nullable) is None
        # and model_fields_set contains the field
        if self.var_field is None and "var_field" in self.model_fields_set:
            _dict['field'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of ErrorDetail from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "code": obj.get("code"),
            "message": obj.get("message"),
            "field": obj.get("field")
        })
        return _obj