# backend/src/app/generated/src/generated_fastapi_server/models/update_user_profile_request.py
# coding: utf-8

"""
    Simple Note Application API

    A simple note-taking application API that allows anonymous browsing of latest public notes  and authenticated access to personal plain text notebook features.  ## Authentication - Public endpoints require no authentication - Private endpoints use Firebase Auth with session cookies - Anonymous users are automatically created for notebook access - Regular users can access basic profile management features 

    The version of the OpenAPI document: 1.0.0
    Contact: support@example.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
from typing_extensions import Annotated
from generated_fastapi_server.models.update_user_profile_request_settings import UpdateUserProfileRequestSettings
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class UpdateUserProfileRequest(BaseModel):
    """
    Request payload for updating user profile
    """ # noqa: E501
    display_name: Optional[Annotated[str, Field(min_length=1, strict=True, max_length=100)]] = Field(default=None, description="Updated display name", alias="displayName")
    avatar_url: Optional[StrictStr] = Field(default=None, description="Updated avatar image URL", alias="avatarUrl")
    settings: Optional[UpdateUserProfileRequestSettings] = None
    __properties: ClassVar[List[str]] = ["displayName", "avatarUrl", "settings"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of UpdateUserProfileRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of settings
        if self.settings:
            _dict['settings'] = self.settings.to_dict()
        # set to None if display_name (nullable) is None
        # and model_fields_set contains the field
        if self.display_name is None and "display_name" in self.model_fields_set:
            _dict['displayName'] = None

        # set to None if avatar_url (nullable) is None
        # and model_fields_set contains the field
        if self.avatar_url is None and "avatar_url" in self.model_fields_set:
            _dict['avatarUrl'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of UpdateUserProfileRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "displayName": obj.get("displayName"),
            "avatarUrl": obj.get("avatarUrl"),
            "settings": UpdateUserProfileRequestSettings.from_dict(obj.get("settings")) if obj.get("settings") is not None else None
        })
        return _obj