# backend/src/app/generated/src/generated_fastapi_server/models/user_profile_response.py
# coding: utf-8

"""
    Simple Note Application API

    A simple note-taking application API that allows anonymous browsing of latest public notes  and authenticated access to personal plain text notebook features.  ## Authentication - Public endpoints require no authentication - Private endpoints use Firebase Auth with session cookies - Anonymous users are automatically created for notebook access - Regular users can access basic profile management features 

    The version of the OpenAPI document: 1.0.0
    Contact: support@example.com
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict
from typing import Any, ClassVar, Dict, List
from generated_fastapi_server.models.api_response_status import ApiResponseStatus
from generated_fastapi_server.models.user_profile import UserProfile
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class UserProfileResponse(BaseModel):
    """
    Response wrapper for user profile
    """ # noqa: E501
    status: ApiResponseStatus
    data: UserProfile
    __properties: ClassVar[List[str]] = ["status", "data"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of UserProfileResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of data
        if self.data:
            _dict['data'] = self.data.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of UserProfileResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "status": obj.get("status"),
            "data": UserProfile.from_dict(obj.get("data")) if obj.get("data") is not None else None
        })
        return _obj